Move applyThousandSeparator to client/utils/index.js file
